mm-machineapp-template example/server/requirements.txt
bottle==0.12.19
websockets==7.0
paho-mqtt==1.5.1
paste==3.5